+++ RenderSystems/GLES2/src/GLSLES/src/OgreGLSLESPreprocessor.cpp
@@ -115,7 +115,7 @@ bool CPreprocessor::Token::GetValue (long &oValue) const
 
     for (; i < Length; i++)
     {
-        long c = long (String [i]);
+        int c = int (String [i]);
         if (isspace (c))
             // Possible end of number
             break;
@@ -174,7 +174,7 @@ int CPreprocessor::Token::CountNL ()
         return 0;
 
     const char *s = String;
-    int l = Length;
+    size_t l = Length;
     int c = 0;
     while (l > 0)
     {
@@ -761,7 +761,7 @@ CPreprocessor::Token CPreprocessor::GetArgument (Token &oArg, bool iExpand)
         }
     }
 
-    uint len = oArg.Length;
+    size_t len = oArg.Length;
     while (true)
     {
         Token t = GetToken (iExpand);
