+++ RenderSystems/GLES2/src/OgreGLES2Texture.cpp
@@ -271,7 +271,7 @@ namespace Ogre {
             GLenum datatype = GLES2PixelUtil::getGLOriginDataType(mFormat);
 
             // Run through this process to pregenerate mipmap pyramid
-            for(size_t mip = 0; mip <= mNumMipmaps; mip++)
+            for(GLint mip = 0; mip <= mNumMipmaps; mip++)
             {
 #if OGRE_DEBUG_MODE
                 LogManager::getSingleton().logMessage("GLES2Texture::create - Mip: " + StringConverter::toString(mip) +
